Clean all project and allow to check with online rules.

+++ Checker/nbt_func.py
@@ -1,14 +1,12 @@
-import traceback
+import traceback,os
 from nbt import nbt
-import config
-from lib.sugar import timer
-from lib.log_color import log,write_log
-import os
-import nbt_rule
+from Checker import config, nbt_rule
+from Checker.lib.sugar import timer
+from Checker.lib.log_color import log,write_log
 from Checker.lib.file_size_io import wait_for_file_transfer_complete
-from Checker.hash_map_handler import calculate_md5
-from Checker.rule_handler import save_md5,load_rule,extract_rules
-
+from Checker.lib.hash_map_handler import calculate_md5
+from Checker.lib.rule_handler import save_md5,load_rule,extract_rules
+from Checker.lib import file_handle
 
 log_directory = 'logs'  # 日志文件夹
 os.makedirs(log_directory, exist_ok=True)  # 创建日志文件夹（如果不存在）
@@ -40,11 +38,18 @@ def count_block_ids(data):
     return block_count,total_count
 
 def check_handler(player_name, filename):
-
+    problem_path = f"save/problem_schematic/{player_name}/{filename}"
     is_cheat_schematic,hash_md5 = main_check(player_name, filename)
     if is_cheat_schematic:
+        # 移动文件到异常蓝图
+        file_handle.move_file(f'save/{filename}', problem_path)
         log.error(f"筛查到异常蓝图: {player_name}/{filename}")
         write_log(f"↑↑↑↑|{hash_md5}|异常蓝图:{player_name}/{filename}")
+        file_handle.copy_file_to_year_folder("rule/chanhuishu.nbt", f"{config.schematics_path}/{player_name}")
+        os.rename(f"{config.schematics_path}/{player_name}/chanhuishu.nbt", f"{config.schematics_path}/{player_name}/{filename}")
+        log.info(f"触发蓝图替换")
+    else:
+        file_handle.move_file(f'save/{filename}', f"{config.schematics_path}/{player_name}/{filename}")
 
 def delete_file(file_path):
     try:
@@ -58,36 +63,48 @@ def delete_file(file_path):
         log.error(f"删除文件时发生错误: {e}")
 
 def main_check(name, file):
+
+    check_result = 1
     dead = False  # whether is cheat schematic
-    blue_print_path = f"{config.schematics_path}/{name}/{file}"
-    try:
-        global_rule = load_rule(convert_to_string=True)
-        block_rule, palette_rule, redundant_rule = extract_rules(global_rule)
-    except Exception as e:
-        log.error(f"Failed to load rule: {e}")
-        global_rule = None
+    source_path = f"{config.schematics_path}/{name}/{file}"
+    check_path = f'save/{file}'
 
-    # log.info("进入检查")
-    hash_trust = load_rule(path="rule/schematics.yml")
-    hash_cal = calculate_md5(blue_print_path)
-    if hash_trust is not None:
-        for hash_item in hash_trust['md5_hashes']:
-            hash_item=hash_item.split("|")
-            if hash_item[0] == hash_cal:
-                if hash_item[1]==str(global_rule.get('version')):
-                    log.info(f"检查到指纹历史:[{name}|{file}]")
-                    # log.debug(hash_item)
-                    if hash_item[2] == "True":
-                        write_log("相同的文件md5，查看历史日志获得蓝图问题")
-                        return True,hash_cal
-                    else:
-                        return False,hash_cal
-    else:
-        delete_file("rule/schematics.yml")
-            # log.debug(hash_item)
-    # log.info(hash_cal)
+    complete = wait_for_file_transfer_complete(source_path)
+    if complete:
+        file_handle.move_file(source_path, check_path)
+        file_handle.copy_file_to_year_folder(check_path, f"save/backup/{name}")
+        blue_print_path = check_path
 
-    complete = wait_for_file_transfer_complete(blue_print_path)
+
+
+        try:
+            global_rule = load_rule(convert_to_string=True)
+            block_rule, palette_rule, redundant_rule = extract_rules(global_rule)
+        except Exception as e:
+            log.error(f"Failed to load rule: {e}")
+            global_rule = None
+
+        # log.info("进入检查")
+        hash_trust = load_rule(path="rule/schematics.yml")
+        hash_cal = calculate_md5(blue_print_path)
+        if hash_trust is not None:
+            for hash_item in hash_trust['md5_hashes']:
+                hash_item=hash_item.split("|")
+                if hash_item[0] == hash_cal:
+                    if hash_item[1]==str(global_rule.get('version')):
+                        log.info(f"检查到指纹历史:[{name}|{file}]")
+                        # log.debug(hash_item)
+                        if hash_item[2] == "True":
+                            write_log("相同的文件md5，查看历史日志获得蓝图问题")
+                            return True,hash_cal
+                        else:
+                            return False,hash_cal
+        else:
+            delete_file("rule/schematics.yml")
+                # log.debug(hash_item)
+        # log.info(hash_cal)
+
+
     if complete:   # 检查文件是否传输完成
         if config.count_block:
             # 统计方块信息
@@ -101,7 +118,7 @@ def main_check(name, file):
         interesting = []
         for rule in global_rule.get('rules'):
             interesting.append(rule.get('block'))
-        str_result = nbt_rule.str_check(blue_print_path,interesting,config.ban_tags,config.ban_block)
+        str_result = nbt_rule.str_check(blue_print_path, interesting, config.ban_tags, config.ban_block)
         if str_result == -1:
             log.error("包含异常标签，蓝图为创造蓝图或篡改蓝图！")
             write_log("包含异常标签，蓝图为创造蓝图或篡改蓝图！")
@@ -112,7 +129,7 @@ def main_check(name, file):
         else:
             try:
 
-                check_result = nbt_rule.rule_check(blue_print_path, block_rule, palette_rule, redundant_rule,nbt_config=global_rule)
+                check_result = nbt_rule.rule_check(blue_print_path, block_rule, palette_rule, redundant_rule, nbt_config=global_rule)
                 if check_result >= 1:
                     log.warning(f"替换规则触发次数： {check_result}")
                 if check_result == -1:
@@ -127,7 +144,8 @@ def main_check(name, file):
                     traceback.print_exc()
 
     path_md5 = r"rule/schematics.yml"
-    save_md5(path_md5,f"{hash_cal}|{global_rule['version']}|{dead}")
+    if check_result < 1:  # 触发替换规则的蓝图不会记录指纹，因为下次再见到仍然需要替换，无法降低耗时
+        save_md5(path_md5,f"{hash_cal}|{global_rule['version']}|{dead}")
     return dead,hash_cal
